backend/requirements.txt
requests
beautifulsoup4
tqdm
fastapi
uvicorn
faiss-cpu
transformers
scikit-learn
python-dotenv
deep_translator
langdetect
numpy
huggingface_hub
fastembed